pyproject.toml
[project]
name = "visual-defect-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "fastapi>=0.136.1",
    "minio>=7.2.20",
    "mlflow>=3.11.1",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "supabase>=2.29.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "ultralytics>=8.4.41",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
